sql: fix in-memory stmt stats not properly updating
Previsouly, in-memeory statement statistics does not return
any metadata. This is due to the temp SQLStats object used
for aggregating RPC fanout results not properly updating the
metadata fields.
This commit address the bug and expanded testing to randomly
test all fields for statement statistics metadata.

Resolves #69363

Release justification: Category 2: Bug fixes and low-risk updates
to new functionality

Release note: None
