+++ pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "json_decoding.go",
         "json_encoding.go",
         "json_impl.go",
+        "testutils.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/sqlstats/persistedsqlstats/sqlstatsutil",
     visibility = ["//visibility:public"],
@@ -15,6 +16,7 @@ go_library(
         "//pkg/util/json",
         "@com_github_cockroachdb_apd_v2//:apd",
         "@com_github_cockroachdb_errors//:errors",
+        "@com_github_stretchr_testify//require",
     ],
 )
